Separated Commands for BodyCam
Separated the commands into a function that is called by the commands. Did this so that we can register exports/events to trigger the bodycam state from other scripts.

I plan to make a plugin script to have the bodycam activate automatically if certain conditions are met. Also added an export to get the current state of the bodycam.